pyproject.toml
[project]
name = "clean-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=3.2,<4",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.135.3",
    "mypy>=1.20.0",
    "passlib>=1.7.4",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.3.3",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "ruff>=0.15.8",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "types-passlib>=1.7.7.20260211",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning"
]
